Update
I added the main(argv) so it can be passed to the command line. The
arguments are just the location of the raw data and the location where
you want the file to be written.

I change the output from Import_5 from tab to cvs. Which makes it
easier to run in R.

I added a column at the end that sum occurrences of that TCR.

The FreqHeatmap.R program now sorts based on the number of occurrences.
I also put in system arguments. args[1] = data, args[2] = number of
rows to analyze, args[3] = number of columns to analyze

I also created a document call standby code which is code that works
for other objectives I have. I took it out of the main Import_5 to
unclutter it.